+++ plugins/http2_adapter/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
+*None.*
+
+## 2.3.2
+
 - Implement `sendTimeout` and `receiveTimeout` for the adapter.
 - Fix redirect not working when requestStream is null.
+- Ignores `Duration.zero` timeouts.
 
 ## 2.3.1+1